Paired icon: Support custom theme.
Usage of 'shell-link' styleclass on paired Icon (which I like and prefer), doesnot work with custom themes as they do not convert all styleclass.
This applies accent color with lighten 20% (if theme is dark) or darken 10% (if theme is light) to the paired Icon in quick setting,
Similar to how shell-link styleclass behaved.

link_color: if($variant == 'light', st-darken(-st-accent-color, 10%), st-lighten(-st-accent-color, 20%));